requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3
requests==2.32.5
python-dotenv==1.2.1

# Google Gemini (Generative AI)
google-generativeai==0.8.5
google-api-core==2.27.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
protobuf==5.29.5
